web-api/Python/requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.2
pydantic~=2.11.3
SQLAlchemy~=2.0.40
pydantic-settings~=2.9.1
email-validator~=2.2.0
tzdata~=2025.2
passlib~=1.7.4
starlette~=0.46.2
python-jose~=3.4.0